Move Unauthenticated App sub components into its own folder
