+++ frontend/src/UnauthenticatedApp/ForgotPasswordPage/ForgotPasswordForm/index.tsx
@@ -1,11 +1,9 @@
-import React from "react";
-import { Formik } from "formik";
-import * as yup from "yup";
-import { toast } from "react-toastify";
-
-import api from "../../api";
-
 import { Button, Grid, Link, TextField } from "@material-ui/core";
+import { Formik } from "formik";
+import React from "react";
+import { toast } from "react-toastify";
+import * as yup from "yup";
+import api from "../../../api";
 
 const validationSchema = yup.object().shape({
   email: yup
